Murs affichés mais mauvaise place, mapX et mapY inversés, incompréhension des directions

+++ includes/cub3d.h
@@ -49,8 +49,8 @@ typedef struct		ray_s
 	double			deltaDistY;
 	double			sideDistX;	
 	double			sideDistY;
-	double			drawStart;
-	double			drawEnd;
+	int				drawStart;
+	int				drawEnd;
 	double			planX;
 	double			planY;
 	double			rDirX;
